+++ .flake8
@@ -4,4 +4,4 @@ inline-quotes = "
 multiline-quotes = """
 docstring-quotes = """
 avoid-escape = False
-ignore = Q000,WPS306,I001,I005,WPS229,D400,WPS317,S101,WPS507,DAR101,DAR201,WPS112,F401,WPS300,WPS412,DAR301,D401,D205,WPS615,I004,WPS110,WPS420,C812
+ignore = Q000,WPS306,I001,I005,WPS229,D400,WPS317,S101,WPS507,DAR101,DAR201,WPS112,F401,WPS300,WPS412,DAR301,D401,D205,WPS615,I004,WPS110,WPS420,C812,WPS404,WPS221,WPS305,WPS323,WPS204,WPS226,WPS220,WPS214
