+++ setup.cfg
@@ -31,4 +31,4 @@ install_requires =
 
 [flake8]
 max-line-length = 120
-extend-ignore = Q000,WPS306,I001,I005,WPS229,D400,WPS317,S101,WPS507,DAR201,DAR101,WPS112,F401,WPS300,WPS412,DAR301,D401,D205,WPS615,I004,WPS110,WPS420,C812
+extend-ignore = Q000,WPS306,I001,I005,WPS229,D400,WPS317,S101,WPS507,DAR201,DAR101,WPS112,F401,WPS300,WPS412,DAR301,D401,D205,WPS615,I004,WPS110,WPS420,C812,WPS404,WPS221,WPS305,WPS323,WPS204,WPS226,WPS220,WPS214
